requirements.txt
# requirements_clean.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file requirements_clean.txt
# platform: linux-64

python=3.10
seaborn=0.12.2
scipy=1.11.3
scikit-learn=1.3.0
rfc3986-validator=0.1.1
rfc3339-validator=0.1.4
requests=2.31.0
cryptography=41.0.3
contourpy=1.0.5
cffi=1.15.1
cycler=0.11.0
decorator=5.1.1
fonttools=4.25.0
idna=3.4
joblib=1.2.0
kiwisolver=1.4.4
matplotlib=3.7.2
numpy=1.26.0
packaging=23.1
pandas=2.1.1
pillow=10.0.1
pycparser=2.21
pyparsing=3.0.9
python-dateutil=2.8.2
pytz=2023.3.post1
six=1.16.0
threadpoolctl
urllib3=1.26.16
pytorch
torchvision
torchaudio
cudatoolkit=11.8
pytorch
